WIP: Authorization
Trying something new with regards to managing permissions.  Now that spicedb is in place, thinking through the best way to authorized endpoints.
